Add files via upload
Registered users normally provide some sort of credentials (such as a username or e-mail address, and a password) to the system in order to prove their identity this is known as logging in. Systems intended for use by the general public often allow any user to register simply by selecting a register or sign up function and providing these credentials for the first time.

This is a developable login and register system. Enter your username and password to register or login to account. Automatically stores the entered account in its memory. When you want to login again, you can enter without any problems. If you try to login when you don't have an account, it will give you an error message. If you already have an account, it will give a successfully login message and than we can design it as we like for the rest of the program.